+++ tests/core/common/engines/alpha/steps/tools.py
@@ -328,7 +328,7 @@ TOOLS: dict[str, dict[str, Any]] = {
     },
     "register_for_confusing_sweepstake": {
         "name": "register_for_confusing_sweepstake",
-        "description": "Register for a sweepstake with more confusing paramater options",
+        "description": "Register for a sweepstake with more confusing parameter options",
         "module_path": "tests.tool_utilities",
         "parameters": {
             "first_name": (
@@ -479,7 +479,7 @@ TOOLS: dict[str, dict[str, Any]] = {
             },
             "pincode": {
                 "type": "string",
-                "description": "the pincode for the account the coins are transfered from",
+                "description": "the pincode for the account the coins are transferred from",
             },
         },
         "required": ["amount", "from_account", "to_account", "pincode"],
@@ -783,6 +783,20 @@ TOOLS: dict[str, dict[str, Any]] = {
         },
         "required": ["boolean"],
     },
+    "check_current_time": {
+        "name": "check_current_time",
+        "description": "Check the current time",
+        "module_path": "tests.tool_utilities",
+        "parameters": {},
+        "required": [],
+    },
+    "check_current_time_emit": {
+        "name": "check_current_time_emit",
+        "description": "Check the current time",
+        "module_path": "tests.tool_utilities",
+        "parameters": {},
+        "required": [],
+    },
 }
 
 
@@ -950,7 +964,7 @@ def given_max_engine_iteration(
     given,
     parsers.parse('a tool relationship whereby "{tool_a}" overlaps with "{tool_b}"'),
 )
-def given_an_overlaping_tools_relationship(
+def given_an_overlapping_tools_relationship(
     context: ContextOfTest,
     tool_a: str,
     tool_b: str,