+++ tests/tool_utilities.py
@@ -487,3 +487,11 @@ def give_boolean_types(
     return ToolResult(
         f"Types for boolean is: {type(boolean[0])} and optional boolean: {type(optional_boolean)}"
     )
+
+
+def check_current_time() -> ToolResult:
+    return ToolResult(data="Current time is 18:03", control={"lifespan": "response"})
+
+
+def check_current_time_emit() -> ToolResult:
+    return ToolResult(data="Current time is 9:59", control={"lifespan": "session"})